Making sure that MainModel doesnt depend on JsonHandler
